// src/types/generated/IntentAnalysisHistory.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DecisionQAPair } from "./DecisionQAPair";
import type { OpeningIntent } from "./OpeningIntent";

export interface IntentAnalysisHistory { id: bigint, sessionFilePath: string, qaPairs: Array<DecisionQAPair>, openingIntent: OpeningIntent, language: string, analyzedAt: string, createdAt: string, }